configured yay to only be installed if it's not already installed

+++ install/os_arch/install_yay.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 
 
+if command -v "yay" &> /dev/null
+then
+    exit
+fi
+
 # download yay
 sudo rm -r $SRC/yay 2> /dev/null
 git clone https://aur.archlinux.org/yay.git $SRC/yay
